+++ src/election_data_analysis/juris/__init__.py
@@ -1,12 +1,14 @@
+import os
 import os.path
 
-from election_data_analysis import database as db
+from election_data_analysis import (
+    database as db,
+    munge as m,
+    userinterface as ui,
+)
 import pandas as pd
 from pandas.api.types import is_numeric_dtype
 from typing import Optional, Dict, Any, List
-from election_data_analysis import munge as m
-from election_data_analysis import user_interface as ui
-from election_data_analysis import preparation as prep
 import numpy as np
 from pathlib import Path
 import csv
@@ -126,7 +128,7 @@ def ensure_juris_files(juris_path: str, ignore_empty: bool = False) -> Optional[
 
     project_root = Path(__file__).parents[1].absolute()
     templates_dir = os.path.join(
-        project_root, "juris_and_munger", "jurisdiction_templates"
+        project_root, "juris", "jurisdiction_templates"
     )
     # notify user of any extraneous files
     extraneous = [
@@ -285,7 +287,7 @@ def find_ambiguous_names(element: str, cf_path: str) -> List[str]:
 
 def check_ru_file(juris_path: str, juris_true_name: str) -> Optional[dict]:
     err = None
-    ru = prep.get_element(juris_path, "ReportingUnit")
+    ru = get_element(juris_path,"ReportingUnit")
 
     # create set of all parents, all lead rus
     parents = set()
@@ -694,4 +696,82 @@ def load_or_update_contests(
             f"Contests not loaded to database (sql error {ie}). "
             f"Check CandidateContest.txt or BallotMeasureContest.txt for errors.",
         )
+    return err
+
+
+def primary(row: pd.Series, party: str, contest_field: str) -> str:
+    try:
+        pr = f"{row[contest_field]} ({party})"
+    except KeyError:
+        pr = None
+    return pr
+
+
+def get_element(juris_path: str, element: str) -> pd.DataFrame:
+    """<juris> is path to jurisdiction directory. Info taken
+    from <element>.txt file in that directory. If file doesn't exist,
+    empty dataframe returned"""
+    f_path = os.path.join(juris_path, f"{element}.txt")
+    if os.path.isfile(f_path):
+        element_df = pd.read_csv(
+            f_path,
+            sep="\t",
+            dtype="object",
+            encoding=jm.default_juris_encoding,
+        )
+    else:
+        element_df = pd.DataFrame()
+    return element_df
+
+
+def remove_empty_lines(df: pd.DataFrame, element: str) -> pd.DataFrame:
+    """return copy of <df> with any contentless lines removed.
+    For dictionary element, such lines may have a first entry (e.g., CandidateContest)"""
+    working = df.copy()
+    # remove all rows with nothing
+    working = working[((working != "") & (working != '""')).any(axis=1)]
+
+    if element == "dictionary":
+        working = working[(working.iloc[:, 1:] != "").any(axis=1)]
+    return working
+
+
+def write_element(
+    juris_path: str, element: str, df: pd.DataFrame, file_name=None
+) -> dict:
+    """<juris> is path to target directory. Info taken
+    from <element>.txt file in that directory.
+    <element>.txt is overwritten with info in <df>"""
+    err = None
+    # set name of target file
+    if not file_name:
+        file_name = f"{element}.txt"
+    # dedupe the input df
+    dupes_df, deduped = ui.find_dupes(df)
+
+    if element == "dictionary":
+        # remove empty lines
+        deduped = remove_empty_lines(deduped, element)
+    try:
+        # write info to file (note: this overwrites existing info in file!)
+        deduped.drop_duplicates().fillna("").to_csv(
+            os.path.join(juris_path, file_name),
+            index=False,
+            sep="\t",
+            encoding=jm.default_juris_encoding,
+        )
+    except Exception as e:
+        err = ui.add_new_error(
+            err,
+            "system",
+            "REMOVEpreparation.write_element",
+            f"Unexpected exception writing to file: {e}",
+        )
+    return err
+
+
+def add_defaults(juris_path: str, juris_template_dir: str, element: str) -> dict:
+    old = get_element(juris_path, element)
+    new = get_element(juris_template_dir, element)
+    err = write_element(juris_path, element, pd.concat([old, new]).drop_duplicates())
     return err
