+++ src/election_data_analysis/juris/jurisdiction_templates/Election.txt
