+++ src/election_data_analysis/juris/jurisdiction_templates/ReportingUnit.txt
